chore: refactor carousel component
